ci: make artifact download more reliable
Yet another attempt to make artifact downloads more reliable.
While it is good that we don't require *every* workflow to be green for it to be selected, the ones with our artifact need to succeed.
With this PR the last three runs are checked, only if those don't work out, we error out